refactor: 커맨드 클래스 패키징 변경
