+++ src/main/java/chess/controller/GameController.java
@@ -13,9 +13,9 @@ import chess.repository.jdbc.JdbcRoomDao;
 import chess.service.GameService;
 import chess.view.InputView;
 import chess.view.OutputView;
-import chess.view.dto.GameCommandType;
-import chess.view.dto.GameRequest;
-import chess.view.dto.MoveRequest;
+import chess.view.dto.game.GameCommandType;
+import chess.view.dto.game.GameRequest;
+import chess.view.dto.game.MoveRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
