+++ src/main/java/chess/controller/UserController.java
@@ -7,8 +7,8 @@ import chess.repository.jdbc.JdbcUserDao;
 import chess.service.UserService;
 import chess.view.InputView;
 import chess.view.OutputView;
-import chess.view.dto.ReadyCommandType;
-import chess.view.dto.ReadyRequest;
+import chess.view.dto.ready.ReadyCommandType;
+import chess.view.dto.ready.ReadyRequest;
 import java.util.List;
 
 public class UserController extends Controller {
