+++ src/main/java/chess/view/dto/ready/ReadyCommandType.java
@@ -1,4 +1,4 @@
-package chess.view.dto;
+package chess.view.dto.ready;
 
 import static chess.view.InputView.INVALID_INPUT_MESSAGE;
